Add support for fast indexing and faster transfer_file processing
Use the previous contents of the index as a sha1 cache, and
modify calculate_sha1() to use it.  This avoid re-reading the
file for every invocation.

Also, create new index as a temp file (with extension '.new')
and atomically rename it to the correct name when done.

Also, avoid processing the transfer_file_lists excessively,
by caching the list (and total transfer file size) in the config.