Use 0 for size of directories
Directories can be different sizes, even when they hold identical
content.  Since the size is used as part of content used to
generate the sha1 for parent directories, this was causing failure
to synchronize for parent dirs up the tree.

Just use 0 as the directory size, instead of the size from the
filesystem, to avoid this problem.