Allow use of simplified configs for sender and receiver
Previously, parts of the code assumed that both sender and receiver
have all config items configured (both source_dir and dest_dir).  This
code separates out references to both of those, depending on the
program mode, so that that configuration (particularly for the
receiver) doesn't need to include attributes that are not needed.

The config files really should be separate anyway.  In the early
releases of this program, a shared config was used for testing,
but in real usage the configs will almost never be the same, because
the sync directory will likely be different on different machines.