Support different path separators for different platforms
Windows uses backslash as os.path.sep.  Convert sneaksync to
always use forward slash (posix_sep) in index and transfer files,
and directory sha1 calculations.  This allows Windows and Linux
systems to interoperate.

Note that it is possible to sync files with backslashes in their
name from Linux to Linux, but not from Linux to Windows.

Also rename  full_path() function to fs_path(), to distinguish it
from full_path variable and make the code clearer.