Add exception handling for various file operations
When moving from just the Picture directory to the full Users/Family
directory, several file access problems were encountered.  These
manifested themselves with python exceptions.

This commit adds exception handling for several cases that cause
problems for sneaksync:
1. junction directories cannot be accessed using normal python filesystem
operations.  These are ignored.
2. files that are open by another program cannot be openened by sneaksync
3. files that this user doesn't have permission to access
4. Some files have UTF16 filenames in Windows, that cause problems for
sneaksync to process on a Linux (UTF8) filesystem.  This problem is still
outstanding.  There should be a way to convert the encoding from UTF16 to
UTF8, but I can't reproduce the problem manually yet.

In all these cases, ignore the error and move on.  In case 4, also print
an error.

Move sneaksync to version 1.0.3