Fix bug with weird characters from Windows
Windows filesystems use mbcs encoding.  Some characters, such as
e with a grave accent, are encoded in a way that makes python
fault on Linux when manipulating these filenames.  This commit
adds support to convert all filenames to utf8.  That is now the
format for both paths in both index files and transfer files.

Also, conversion back to mbcs was added if the receiver is
Windows.  This has not been tested.