Add date and time to index header
This can be used to manually check whether the index files are
getting synced by the cloud drive mechanism.  For example, if you
know the dest_index is being generated, but it's not showing up
at the source, then the cloud synchronization is not working.

This happened to me one time when dropbox didn't start automatically
on a windows system reboot, and syncs didn't occur for a few weeks.